Added a hangman game and some new buttons
Added a hangman game where each person in the chat gets to play a game of hangman. You can start a game by typing /game start or by simply pressing the "Game" button. There is also a "Help" button that outputs the /help text. There is a new funciton called SendPacketToAll which sends a packet to all clients, useful for sending a message to everyone.