Added a new winform for painting plus some improvements
Added a new winform which allows you to paint on it, this is replacing monogame. You can start it with /paint or by pressing the paint button. Pressing any button on the mouse will draw on the canvas, there are 4 sliders representing RBGA and a box at the top to show you what colour you're currently using. It is run on a seperate thread. Updated the help text.